+++ metadata.rb
@@ -1,7 +1,10 @@
-name 'osquery-pkg'
+name 'osquery'
 maintainer 'Jack Naglieri'
-maintainer_email 'naglieri@yahoo-inc.com'
+maintainer_email 'jnaglierijr@ymail.com'
 license 'All rights reserved'
-description 'Installs/Configures osquery-pkg'
+description 'Install and configure osquery'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+supports 'mac_os_x'
 version '0.2.0'
+
+depends 'homebrew'